requirements.txt
bcrypt==4.2.1
mysql-connector-python==9.1.0
pytest==8.3.4
